+++ objectbox/example/flutter/objectbox_demo_sync/android/app/src/main/kotlin/com/example/objectbox_demo/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.example.objectbox_demo
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}