+++ objectbox/CHANGELOG.md
@@ -1,4 +1,23 @@
-## latest
+## 1.5.0 (2022-05-11)
+* Add `Store.runInTransactionAsync` to run database operations asynchronously in the background
+  (requires Flutter 2.8.0/Dart 2.15.0 or newer). #415
+* Rename `Store.runIsolated` to `runAsync`, drop unused `mode` parameter, propagate errors and
+  handle premature isolate exit. #415
+* The native ObjectBox library is also searched for in the `lib` subfolder on desktop OS (macOS,
+  Linux, Windows). This is where the [`install.sh`](/install.sh) script downloads it by default.
+  E.g. it is no longer necessary to install the library globally to run `dart test` or `flutter test`.
+* Windows: Support database directory paths that contain unicode (UTF-8) characters. #406
+* Changed `Query.stream` to collect results in a worker isolate, which should typically be faster. #420
+* Update: [objectbox-c 0.16.0](https://github.com/objectbox/objectbox-c/releases/tag/v0.16.0).
+* Update: [objectbox-android 3.1.3](https://github.com/objectbox/objectbox-java/releases/tag/V3.1.3).
+* Add new [task with tag list Flutter example app](example/flutter/objectbox_demo_relations) that 
+  shows how to use relations. #419
+
+## 1.4.1 (2022-03-01)
+
+* Resolve "another store is still open" issue after Flutter hot restart (hot reload continues to work). #387
+* Add `Store.isClosed()`. #390
+* Add note to `objectbox.g.dart` on how to re-generate (update) it.
 
 ## 1.4.0 (2022-02-22)